Implemented cost for FileScanNode and estimateBoolOperSelectivity
Calculated selectivity (using the helper function
estimateSelectivity for when we have a predicate). Use selectivity
to estimate the number of tuples produced and calculate cost
based on this number.

Calculated the selectivity for AND_EXPR, OR_EXPR, and NOT_EXPR
by iterating through the boolean terms for AND and OR and
updating the selectivity accordingly.