Implemented BTreeTupleFile and LeafPageOperations
Implemented the helper method navigateToLeafPage in
BTreeTupleFile which navigates from the root of the
Btree to the appropriate leaf node based on the search
key provided by the caller.

Partially implemented the helper method
splitLeafAndAddTuple which splits the specified leaf
node into two nodes and updates the parent node and
inserts the specified tuple into the appropriate leaf.