movePointers for Inner Nodes and IndexHandling
Implemented movePointersLeft and movePointers Right
within the innerPage class. These methods will
redistribute a specified amount of pointers and their
corresponding tuples to a sibling node.

Implemented addRowToIndexes and removeRowFromIndexes
which are called after there is an update to a table
with indexes. THese methods update the corresponding
index file for each table.